split utilities into proper modules
