+++ crates/bevy_reflect/derive/Cargo.toml
@@ -20,10 +20,10 @@ functions = []
 
 [dependencies]
 bevy_macro_utils = { path = "../../bevy_macro_utils", version = "0.15.0-dev" }
-
-syn = { version = "2.0", features = ["full"] }
+perfect-derive = "0.1.3"
 proc-macro2 = "1.0"
 quote = "1.0"
+syn = { version = "2.0", features = ["full"] }
 uuid = { version = "1.1", features = ["v4"] }
 
 [lints]
