+++ crates/bevy_reflect/derive/src/from_reflect.rs
@@ -2,7 +2,8 @@ use crate::container_attributes::REFLECT_DEFAULT;
 use crate::derive_data::ReflectEnum;
 use crate::enum_utility::{EnumVariantOutputData, FromReflectVariantBuilder, VariantBuilder};
 use crate::field_attributes::DefaultBehavior;
-use crate::utility::{ident_or_index, WhereClauseOptions};
+use crate::ident::ident_or_index;
+use crate::where_clause_options::WhereClauseOptions;
 use crate::{ReflectMeta, ReflectStruct};
 use bevy_macro_utils::fq_std::{FQClone, FQDefault, FQOption};
 use proc_macro2::Span;
