+++ crates/bevy_reflect/derive/src/impls/typed.rs
@@ -1,17 +1,16 @@
-use crate::utility::{StringExpr, WhereClauseOptions};
+use crate::derive_data::{ReflectMeta, ReflectTypePath};
+use crate::string_expression::StringExpression;
+use crate::where_clause_options::WhereClauseOptions;
+use bevy_macro_utils::fq_std::FQOption;
+use proc_macro2::TokenStream;
 use quote::{quote, ToTokens};
-
-use crate::{
-    derive_data::{ReflectMeta, ReflectTypePath},
-    utility::wrap_in_option,
-};
 
 /// Returns an expression for a `NonGenericTypeCell` or `GenericTypeCell`  to generate `'static` references.
 fn static_type_cell(
     meta: &ReflectMeta,
     property: TypedProperty,
-    generator: proc_macro2::TokenStream,
-) -> proc_macro2::TokenStream {
+    generator: TokenStream,
+) -> TokenStream {
     let bevy_reflect_path = meta.bevy_reflect_path();
     if meta.type_path().impl_is_generic() {
         let cell_type = match property {
@@ -48,11 +47,11 @@ pub(crate) enum TypedProperty {
     TypePath,
 }
 
-pub(crate) fn impl_type_path(meta: &ReflectMeta) -> proc_macro2::TokenStream {
+pub(crate) fn impl_type_path(meta: &ReflectMeta) -> TokenStream {
     let where_clause_options = WhereClauseOptions::new(meta);
 
     if !meta.attrs().type_path_attrs().should_auto_derive() {
-        return proc_macro2::TokenStream::new();
+        return TokenStream::new();
     }
 
     let type_path = meta.type_path();
@@ -80,9 +79,9 @@ pub(crate) fn impl_type_path(meta: &ReflectMeta) -> proc_macro2::TokenStream {
         )
     };
 
-    let type_ident = wrap_in_option(type_path.type_ident().map(StringExpr::into_borrowed));
-    let module_path = wrap_in_option(type_path.module_path().map(StringExpr::into_borrowed));
-    let crate_name = wrap_in_option(type_path.crate_name().map(StringExpr::into_borrowed));
+    let type_ident = wrap_in_option(type_path.type_ident().map(StringExpression::into_borrowed));
+    let module_path = wrap_in_option(type_path.module_path().map(StringExpression::into_borrowed));
+    let crate_name = wrap_in_option(type_path.crate_name().map(StringExpression::into_borrowed));
 
     let primitive_assert = if let ReflectTypePath::Primitive(_) = type_path {
         Some(quote! {
@@ -132,8 +131,8 @@ pub(crate) fn impl_type_path(meta: &ReflectMeta) -> proc_macro2::TokenStream {
 pub(crate) fn impl_typed(
     meta: &ReflectMeta,
     where_clause_options: &WhereClauseOptions,
-    type_info_generator: proc_macro2::TokenStream,
-) -> proc_macro2::TokenStream {
+    type_info_generator: TokenStream,
+) -> TokenStream {
     let type_path = meta.type_path();
     let bevy_reflect_path = meta.bevy_reflect_path();
 
@@ -149,5 +148,17 @@ pub(crate) fn impl_typed(
                 #type_info_cell
             }
         }
+    }
+}
+
+/// Turns an `Option<TokenStream>` into a `TokenStream` for an `Option`.
+fn wrap_in_option(tokens: Option<TokenStream>) -> TokenStream {
+    match tokens {
+        Some(tokens) => quote! {
+            #FQOption::Some(#tokens)
+        },
+        None => quote! {
+            #FQOption::None
+        },
     }
 }
