+++ crates/bevy_reflect/derive/src/meta.rs
@@ -1,0 +1,7 @@
+use bevy_macro_utils::BevyManifest;
+use syn::Path;
+
+/// Returns the correct path for `bevy_reflect`.
+pub(crate) fn get_bevy_reflect_path() -> Path {
+    BevyManifest::get_path_direct("bevy_reflect")
+}
